Paper-Reader-Pro/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
openai==1.51.0
google-generativeai==0.8.1
pydantic==2.9.2
python-dotenv==1.0.1
